web: break apart OverviewResourceBar
